Move instrument name validation tests into BuilderTests
Review feedback: keep a 1:1 file/test reference. The validation logic now
lives in InstrumentBuilder.swift, so its tests belong with the other
InstrumentBuilder tests rather than in a file named after a type that no
longer exists.

InstrumentNameValidationTests.swift is removed and its four tests plus
the SDK-meter helper move into BuilderTests. The two name syntax tests
are renamed to mention instrument names, since they now sit among tests
for the whole builder family. No assertion changed; trailing commas,
comment alignment and the defer body were reflowed to satisfy swiftformat.
